Share complete platform resolution rule
